chatbot/requirement.txt
Flask
google-generativeai
flask-cors
dot-env